requirements.txt
Flask>=2.2
Werkzeug==2.3.7
Pillow==9.5.0
gunicorn==20.1.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
typing-extensions<4.6.0,>=3.6.6
google-generativeai==0.3.1
pandas==2.0.3 